+++ src/Selenium2Library/locators/elementfinder.py
@@ -1,6 +1,7 @@
 from Selenium2Library import utils
 from robot.api import logger
 from robot.utils import NormalizedDict
+from robot.libraries.BuiltIn import BuiltIn
 
 
 class ElementFinder(object):
@@ -22,6 +23,7 @@ class ElementFinder(object):
             'default': self._find_by_default
         }
         self._strategies = NormalizedDict(initial=strategies, caseless=True, spaceless=True)
+        self._default_strategies = strategies.keys()
 
     def find(self, browser, locator, tag=None):
         assert browser is not None
@@ -34,6 +36,30 @@ class ElementFinder(object):
             raise ValueError("Element locator with prefix '" + prefix + "' is not supported")
         (tag, constraints) = self._get_tag_and_constraints(tag)
         return strategy(browser, criteria, tag, constraints)
+
+    def register(self, strategy, persist):
+        if strategy.name in self._strategies:
+            raise AttributeError("The custom locator '" + strategy.name +
+            "' cannot be registered. A locator of that name already exists.")
+        self._strategies[strategy.name] = strategy.find
+
+        if not persist:
+            # Unregister after current scope ends
+            suite = BuiltIn().get_variable_value('${SUITE NAME}')
+            test  = BuiltIn().get_variable_value('${TEST NAME}', '')
+            scope =  suite + '.' + test if test != '' else suite
+            utils.events.on('scope_end', scope, self.unregister, strategy.name)
+
+    def unregister(self, strategy_name):
+        if strategy_name in self._default_strategies:
+            raise AttributeError("Cannot unregister the default strategy '" + strategy_name + "'")
+        elif strategy_name not in self._strategies:
+            logger.info("Cannot unregister the non-registered strategy '" + strategy_name + "'")
+        else:
+            del self._strategies[strategy_name]
+
+    def has_strategy(self, strategy_name):
+        return strategy_name in self.strategies
 
     # Strategy routines, private
 
@@ -90,7 +116,7 @@ class ElementFinder(object):
         return self._filter_elements(
             browser.find_elements_by_tag_name(criteria),
             tag, constraints)
-            
+
     def _find_by_sc_locator(self, browser, criteria, tag, constraints):
         js = "return isc.AutoTest.getElement('%s')" % criteria.replace("'", "\\'")
         return self._filter_elements([browser.execute_script(js)], tag, constraints)