+++ src/Selenium2Library/utils/events/event.py
@@ -1,0 +1,7 @@
+import abc
+
+class Event(object):
+
+    @abc.abstractmethod
+    def trigger(self, *args, **kwargs):
+        pass