+++ CHANGES.rst
@@ -3,6 +3,11 @@
 
 1.7 (unreleased)
 ----------------
+- Fixed ‘NoSuchWindowException' issue. Running keyword 'Select Window' after 'Close Window'
+  will trigger this issue if locator has prefix 'name=','title=' or 'url='. Also fixed same
+  issue for keywords 'Get Window Ids', 'Get Window Titles' and 'Get Window Names'.
+  [divfor]
+  
 - Corrected error message in new keyword 'Wait Until Element Is Not
   Visible' to reflect element being visible instead of not visible.
   [joepurdy]
@@ -16,6 +21,9 @@
 - Added new keyword 'Element Should Not Contain'.
   [molsky]
 
+- Added new keyword 'Wait Until Page Does Not Contain Element'.
+  [molsky]
+
 - Added new locator strategy, scLocator, for finding SmartClient and SmartGWT elements.
   [IlfirinPL]
 
@@ -25,6 +33,7 @@
 - 'Capture Screenshot' now attempts to create its containing directory if the directory
   specified in the filename does not exist.
 - 'Choose File' now fails if the file doesn't exist
+- Added new keywords 'Add Location Strategy' and 'Remove Location Strategy'
   [zephraph]
 
 - Added 'Get Window Position' and 'Set Window Position' keywords matching the